Split OpenSSL backend into separate files (hashes)
